requirements.txt
python-telegram-bot==21.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.4.0
python-dotenv==1.0.1
supabase==2.24.0
httpx==0.28.1
aiohttp==3.11.11
